zip: separate read logic from data (#9)
Removes the Parser type and replaces it with the Info struct and the Directory struct.  This separates the data that is needed for each logical operation.

This library performs 2 operations to read information about the zip archive.  The first one gathers global information about the archive and the second one reads information specific to all the files.  Rather than have a single type that contains all this information, I've seperated each data set into it's own type and only given access to the information each operation needs.  This separation makes the library more flexible, for example, a program may want to read the global archive information without reading the file information, with the new interface they know exactly what data will be populated by this single operation and also don't need to provide an allocator.